requirements.txt
fastapi[standard]
# psycopg2
psycopg2-binary
python-dotenv
pydantic-settings
uvicorn
langchain
langchain[google-genai]
langchain-community
langchain-docling
langchain-postgres